In a multi-IO-thread scenario, some client checks were moved to IO threads to avoid the main thread pausing every time during clientsCron.

When the number of clients is set to 10000, there is about 20% change.

https://github.com/redis/redis/issues/13885

Here are the test results:

memtier_benchmark --data-size 521 --ratio 1:1 --key-pattern R:R --key-minimum=1 --key-maximum 3000000 --test-time 60 -c 500 -t 20 --hide-histogram -x 3

**After the modification.**

```
BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       228872.62          ---          ---        21.84701        18.43100        94.20700       178.17500    127095.62 
Gets       228789.87       166.42    228623.45        21.83361        18.43100        94.20700       180.22300      8944.73 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     457662.49       166.42    228623.45        21.84031        18.43100        94.20700       179.19900    136040.34 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       223669.53          ---          ---        22.34478        18.68700       105.47100       178.17500    124206.42 
Gets       223587.09       166.52    223420.57        22.35959        18.68700       105.47100       179.19900      8743.96 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     447256.62       166.52    223420.57        22.35219        18.68700       105.47100       178.17500    132950.38 


AGGREGATED AVERAGE RESULTS (3 runs)
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       226230.28          ---          ---        22.10704        18.43100       103.93500       195.58300    125628.37 
Gets       226147.48       166.44    225981.04        22.08480        18.43100       103.93500       196.60700      8842.73 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     452377.76       166.44    225981.04        22.09592        18.43100       103.93500       196.60700    134471.10
```

**Before the modification.**

```
BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       176179.35          ---          ---        28.43698        24.44700       120.31900       201.72700     97831.41 
Gets       176096.14      1011.36    175084.77        28.31696        24.19100       119.80700       204.79900      7336.40 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     352275.48      1011.36    175084.77        28.37698        24.31900       119.80700       202.75100    105167.80 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       167608.62          ---          ---        30.26107        23.42300       199.67900       354.30300     93071.39 
Gets       167523.58       902.81    166620.77        30.13004        23.29500       199.67900       354.30300      6948.66 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     335132.20       902.81    166620.77        30.19558        23.42300       199.67900       354.30300    100020.06 


AGGREGATED AVERAGE RESULTS (3 runs)
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       172947.05          ---          ---        29.05879        24.19100       141.31100       296.95900     96036.28 
Gets       172863.63       973.35    171890.28        28.97310        24.06300       141.31100       299.00700      7191.68 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     345810.68       973.35    171890.28        29.01596        24.19100       141.31100       299.00700    103227.96
```